db queries every 10 shows

+++ backend/shows/tasks.py
@@ -94,12 +94,11 @@ def update_all_shows_task(start_index):
     shows = Show.objects.all()[start_index:]
     count = len(shows)
     i = 1
+    episodes_to_create = []
+    episodes_to_update = []
+    episodes_to_delete_pks = []
 
     for show in shows:
-        episodes_to_create = []
-        episodes_to_update = []
-        episodes_to_delete_pks = []
-
         tmdb_show_id = show.tmdb_id
         key = get_tmdb_show_key(tmdb_show_id)
         try:
@@ -140,10 +139,15 @@ def update_all_shows_task(start_index):
             episodes_to_delete_pks += temp3
 
         print(f'updated {i} of {count}')
-        i += 1
 
-        Episode.objects.filter(pk__in=episodes_to_delete_pks).delete()
-        Episode.objects.bulk_update(episodes_to_update,
-                                    ['tmdb_episode_number', 'tmdb_season', 'tmdb_name', 'tmdb_release_date',
-                                     'tmdb_runtime'])
-        Episode.objects.bulk_create(episodes_to_create)
+        if i % 10 == 0:
+            Episode.objects.filter(pk__in=episodes_to_delete_pks).delete()
+            Episode.objects.bulk_update(episodes_to_update,
+                                        ['tmdb_episode_number', 'tmdb_season', 'tmdb_name', 'tmdb_release_date',
+                                         'tmdb_runtime'])
+            Episode.objects.bulk_create(episodes_to_create)
+            episodes_to_create = []
+            episodes_to_update = []
+            episodes_to_delete_pks = []
+
+        i += 1
